Finish all functionality. Now to test

+++ api/amenity_controller.py
@@ -1,6 +1,8 @@
 import sys
 import os
 
+from flask_jwt_extended import get_jwt_identity, jwt_required
+
 sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
 
 from flask import request, jsonify, Blueprint
@@ -9,12 +11,17 @@ from models.amenity import Amenity
 
 amenity_bp = Blueprint("amenity", __name__)
 
-@amenity_bp.route('/amenities', methods=['GET'])
+@amenity_bp.route('/amenities', methods=['GET'], endpoint="func1")
 def get_all_amenities():
     return DataManager.all(Amenity)
 
-@amenity_bp.route('/amenities', methods=['POST'])
+@amenity_bp.route('/amenities', methods=['POST'], endpoint="func2")
+@jwt_required
 def create_amenity():
+    curr_user = get_jwt_identity()
+    if not curr_user.is_admin:
+        return jsonify("User does not have admin privileges"), 403
+
     jData = request.get_json()
     for field in ["name", "description"]:
         if not isinstance(jData[field], str):
@@ -31,14 +38,14 @@ def create_amenity():
     )
     return DataManager.save(amenity), 201
 
-@amenity_bp.route('/amenities/<int:amenity_id>', methods=['GET'])
+@amenity_bp.route('/amenities/<int:amenity_id>', methods=['GET'], endpoint="func3")
 def get_amenity(amenity_id):
     amenity = DataManager.get(amenity_id, Amenity)
     if not amenity:
         return jsonify({"error": "Amenity not found"}), 404
     return jsonify(amenity), 200
 
-@amenity_bp.route('/amenities/<int:amenity_id>', methods=['PUT'])
+@amenity_bp.route('/amenities/<int:amenity_id>', methods=['PUT'], endpoint="func4")
 def update_amenity(amenity_id):
     jData = request.get_json()
     req_amenity = DataManager.get(amenity_id, Amenity)
@@ -58,7 +65,12 @@ def update_amenity(amenity_id):
     except Exception:
         return "Bad Request", 400
 
-@amenity_bp.route('/amenities/<int:amenity_id>', methods=['DELETE'])
+@amenity_bp.route('/amenities/<int:amenity_id>', methods=['DELETE'], endpoint="func5")
+@jwt_required
 def delete_amenity(amenity_id):
+    curr_user = get_jwt_identity()
+    if not curr_user.is_admin:
+        return jsonify("User does not have admin privileges"), 403
+
     DataManager.delete(amenity_id, Amenity)
     return jsonify("Deleted"), 204
