+++ api/city_controller.py
@@ -1,6 +1,8 @@
 import sys
 import os
 
+from flask_jwt_extended import get_jwt_identity, jwt_required
+
 sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
 
 from flask import request, jsonify, Blueprint
@@ -10,12 +12,17 @@ from models.city import City
 
 city_bp = Blueprint("city_bp", __name__)
 
-@city_bp.route('/cities', methods=['GET'])
+@city_bp.route('/cities', methods=['GET'], endpoint="get_all_cities")
 def get_all_cities():
     return DataManager.all(City)
 
-@city_bp.route('/cities', methods=['POST'])
+@city_bp.route('/cities', methods=['POST'], endpoint="create_city")
+@jwt_required
 def create_city():
+    curr_user = get_jwt_identity()
+    if not curr_user.is_admin:
+        return jsonify("User does not have admin privileges"), 403
+
     jData = request.get_json()
     if not isinstance(jData["name"], str) or not isinstance(jData["population"], int):
         return jsonify("Bad Request"), 400
@@ -33,12 +40,12 @@ def create_city():
     )
     return DataManager.save(city), 201
 
-@city_bp.route('/cities/<int:city_id>', methods=['GET'])
+@city_bp.route('/cities/<int:city_id>', methods=['GET'], endpoint="get_city")
 def get_city(city_id):
     curr_city = DataManager.get(city_id, City)
     return { "name": curr_city["name"], "country_code": curr_city["country_code"] }
 
-@city_bp.route('/cities/<int:city_id>', methods=['PUT'])
+@city_bp.route('/cities/<int:city_id>', methods=['PUT'], endpoint="update_city")
 def update_city(city_id):
     jData = request.get_json()
     req_city = DataManager.get(city_id, City)
@@ -59,8 +66,13 @@ def update_city(city_id):
         return "Bad Request", 400
 
 
-@city_bp.route('/cities/<int:city_id>', methods=['DELETE'])
-def delete_cities(city_id):
+@city_bp.route('/cities/<int:city_id>', methods=['DELETE'], endpoint="delete_city")
+@jwt_required
+def delete_city(city_id):
+    curr_user = get_jwt_identity()
+    if not curr_user.is_admin:
+        return jsonify("User does not have admin privileges"), 403
+
     DataManager.delete(city_id, City)
     return jsonify("Deleted"), 204
